merge from upstream

+++ tests/test_pipeline.py
@@ -1,6 +1,7 @@
 """Test individual components of the analysis pipeline.
 """
 import os
+import sys
 import shutil
 import subprocess
 import unittest
@@ -9,22 +10,30 @@ from nose.plugins.attrib import attr
 
 from bcbio import utils
 from bcbio.bam import fastq
-from bcbio.distributed.messaging import parallel_runner
+# Be back compatible with 0.7.6 -- remove after 0.7.7 release
+try:
+    from bcbio.distributed import prun
+except ImportError:
+    prun = None
 from bcbio.pipeline.config_utils import load_config
 from bcbio.provenance import programs
 from bcbio.variation import vcfutils
+
+sys.path.append(os.path.dirname(__file__))
+from test_automated_analysis import get_post_process_yaml, make_workdir
 
 class RunInfoTest(unittest.TestCase):
     def setUp(self):
         self.data_dir = os.path.join(os.path.dirname(__file__), "data")
 
     @attr(speed=1)
+    @attr(blah=True)
     def test_programs(self):
         """Identify programs and versions used in analysis.
         """
-        config = load_config(os.path.join(self.data_dir, "automated",
-                                          "post_process-sample.yaml"))
-        print programs._get_versions(config)
+        with make_workdir() as workdir:
+            config = load_config(get_post_process_yaml(self.data_dir, workdir))
+            print programs._get_versions(config)
 
 class VCFUtilTest(unittest.TestCase):
     """Test various utilities for dealing with VCF files.
@@ -39,17 +48,21 @@ class VCFUtilTest(unittest.TestCase):
     def test_1_parallel_vcf_combine(self):
         """Parallel combination of VCF files, split by chromosome.
         """
+        # Be back compatible with 0.7.6 -- remove after 0.7.7 release
+        if prun is None:
+            return
         files = [os.path.join(self.var_dir, "S1-variants.vcf"), os.path.join(self.var_dir, "S2-variants.vcf")]
         ref_file = os.path.join(self.data_dir, "genomes", "hg19", "seq", "hg19.fa")
-        config = load_config(os.path.join(self.data_dir, "automated",
-                                          "post_process-sample.yaml"))
-        run_parallel = parallel_runner({"type": "local", "cores": 1}, {}, config)
+        with make_workdir() as workdir:
+            config = load_config(get_post_process_yaml(self.data_dir, workdir))
+            config["algorithm"] = {}
         region_dir = os.path.join(self.var_dir, "S1_S2-combined-regions")
         if os.path.exists(region_dir):
             shutil.rmtree(region_dir)
         if os.path.exists(self.combo_file):
             os.remove(self.combo_file)
-        vcfutils.parallel_combine_variants(files, self.combo_file, ref_file, config, run_parallel)
+        with prun.start({"type": "local", "cores": 1}, [[config]], config) as run_parallel:
+            vcfutils.parallel_combine_variants(files, self.combo_file, ref_file, config, run_parallel)
         for fname in files:
             if os.path.exists(fname + ".gz"):
                 subprocess.check_call(["gunzip", fname + ".gz"])
@@ -61,9 +74,13 @@ class VCFUtilTest(unittest.TestCase):
     def test_2_vcf_exclusion(self):
         """Exclude samples from VCF files.
         """
+        # Be back compatible with 0.7.6 -- remove after 0.7.7 release
+        if prun is None:
+            return
         ref_file = os.path.join(self.data_dir, "genomes", "hg19", "seq", "hg19.fa")
-        config = load_config(os.path.join(self.data_dir, "automated",
-                                          "post_process-sample.yaml"))
+        with make_workdir() as workdir:
+            config = load_config(get_post_process_yaml(self.data_dir, workdir))
+            config["algorithm"] = {}
         out_file = utils.append_stem(self.combo_file, "-exclude")
         to_exclude = ["S1"]
         if os.path.exists(out_file):
@@ -84,7 +101,6 @@ class VCFUtilTest(unittest.TestCase):
                           "/path/2/input/D1HJVACXX_2_AAGAGATC_2.fastq"], out[0]
         assert out[1] == ["/path/to/input/D1HJVACXX_3_AAGAGATC_1.fastq",
                           "/path/2/input/D1HJVACXX_3_AAGAGATC_2.fastq"], out[1]
-
 
         test_pairs = ["/path/to/input/Tester_1_fastq.txt",
                       "/path/to/input/Tester_2_fastq.txt"]
